+++ packages/plugins/presets/guiding-words/vite.config.ts
@@ -1,4 +1,4 @@
-import { vettaPluginFederation } from "@vetta/plugin-vite";
+import { vettaPluginFederation } from "@vetta-org/plugin-vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({